change P_W and P_H to doubles

+++ src/libs/functions/IMinterpreter.cpp
@@ -44,7 +44,7 @@
  ----------------------------------------------------------------------------------
  */
 
-int IMinterpreter(bool* heaterStates,float *P_H,float *P_W){
+int IMinterpreter(bool* heaterStates,double *P_H,double *P_W){
     
     /*
      Inputs/Outputs:
